Add support for "file" outputType
Currently the module only support "blob" or "base64" for outputType.
This commit adds support for a 3rd option: "file" which will return a File javascript object.

This is useful for resizing images that are meant to be uploaded as files. I.e this can be used to send a file to a flask app that uses "FileStorage" class.