Fix bug "Uncaught TypeError: Cannot read property 'name' of undefined"
Fixes a small bug that throws error "index.js:1 Uncaught TypeError: Cannot read property 'name' of undefined" when using the "file" outputType parameter.

This bug is happens because the variable "file" is being overwritten when creating a new file using "const file = xxx".

So just change the name of this variable to something else to get rid of the error and to not overwrite the original file variable used in the code.